requirements.txt
torch~=1.13
torchvision~=0.14
Pillow
numpy